+++ src/main/jchess-web/components/GameComponent/Chessboard/default_chessboard.tsx
