+++ src/main/jchess-web/components/GameComponent/Chessboard/hex_chessboard.tsx
