+++ src/main/jchess-web/components/GameComponent/Chessboard/hex_tile.tsx
